Wallpaper: use JPEG2000-encoded wallpaper
If we extract the wallpaper bit stream from the WdS handout we get JPEG2000.
It’s much better to preserve the original encoding than to re-encode.

Unfortunately the wallpaper now has inverted colors which we can remedy with a
"secret" \includegraphics option: "decodearray={1 0 }". To apply it to the
wallpaper we need to adjust \ThisCenterWallPaper to pass an optional parameter
to \includegraphics.